Refactor: Split ord_dep_graph in hpp and cpp files
