--- include/vectorclock_types.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-typedef int VCValueT;
